+++ test/test-barrier.c
@@ -27,20 +27,22 @@
 
 typedef struct {
   uv_barrier_t barrier;
-  int delay;
-  volatile int posted;
-  int main_barrier_wait_rval;
-  int worker_barrier_wait_rval;
+  unsigned delay;
+  unsigned niter;
+  unsigned main_barrier_wait_rval;
+  unsigned worker_barrier_wait_rval;
 } worker_config;
 
 
 static void worker(void* arg) {
   worker_config* c = arg;
+  unsigned i;
 
   if (c->delay)
     uv_sleep(c->delay);
 
-  c->worker_barrier_wait_rval = uv_barrier_wait(&c->barrier);
+  for (i = 0; i <= c->niter; i++)
+      c->worker_barrier_wait_rval += uv_barrier_wait(&c->barrier);
 }
 
 
@@ -50,16 +52,16 @@ TEST_IMPL(barrier_1) {
 
   memset(&wc, 0, sizeof(wc));
 
-  ASSERT(0 == uv_barrier_init(&wc.barrier, 2));
-  ASSERT(0 == uv_thread_create(&thread, worker, &wc));
+  ASSERT_EQ(0, uv_barrier_init(&wc.barrier, 2));
+  ASSERT_EQ(0, uv_thread_create(&thread, worker, &wc));
 
   uv_sleep(100);
   wc.main_barrier_wait_rval = uv_barrier_wait(&wc.barrier);
 
-  ASSERT(0 == uv_thread_join(&thread));
+  ASSERT_EQ(0, uv_thread_join(&thread));
   uv_barrier_destroy(&wc.barrier);
 
-  ASSERT(1 == (wc.main_barrier_wait_rval ^ wc.worker_barrier_wait_rval));
+  ASSERT_EQ(1, (wc.main_barrier_wait_rval ^ wc.worker_barrier_wait_rval));
 
   return 0;
 }
@@ -72,15 +74,15 @@ TEST_IMPL(barrier_2) {
   memset(&wc, 0, sizeof(wc));
   wc.delay = 100;
 
-  ASSERT(0 == uv_barrier_init(&wc.barrier, 2));
-  ASSERT(0 == uv_thread_create(&thread, worker, &wc));
+  ASSERT_EQ(0, uv_barrier_init(&wc.barrier, 2));
+  ASSERT_EQ(0, uv_thread_create(&thread, worker, &wc));
 
   wc.main_barrier_wait_rval = uv_barrier_wait(&wc.barrier);
 
-  ASSERT(0 == uv_thread_join(&thread));
+  ASSERT_EQ(0, uv_thread_join(&thread));
   uv_barrier_destroy(&wc.barrier);
 
-  ASSERT(1 == (wc.main_barrier_wait_rval ^ wc.worker_barrier_wait_rval));
+  ASSERT_EQ(1, (wc.main_barrier_wait_rval ^ wc.worker_barrier_wait_rval));
 
   return 0;
 }
@@ -89,26 +91,32 @@ TEST_IMPL(barrier_2) {
 TEST_IMPL(barrier_3) {
   uv_thread_t thread;
   worker_config wc;
+  unsigned i;
 
   memset(&wc, 0, sizeof(wc));
+  wc.niter = 5;
 
-  ASSERT(0 == uv_barrier_init(&wc.barrier, 2));
-  ASSERT(0 == uv_thread_create(&thread, worker, &wc));
+  ASSERT_EQ(0, uv_barrier_init(&wc.barrier, 2));
+  ASSERT_EQ(0, uv_thread_create(&thread, worker, &wc));
 
-  wc.main_barrier_wait_rval = uv_barrier_wait(&wc.barrier);
+  for (i = 0; i <= wc.niter; i++)
+      wc.main_barrier_wait_rval += uv_barrier_wait(&wc.barrier);
 
-  ASSERT(0 == uv_thread_join(&thread));
+  ASSERT_EQ(0, uv_thread_join(&thread));
   uv_barrier_destroy(&wc.barrier);
 
-  ASSERT(1 == (wc.main_barrier_wait_rval ^ wc.worker_barrier_wait_rval));
+  ASSERT_EQ(wc.niter + 1, wc.main_barrier_wait_rval + wc.worker_barrier_wait_rval);
 
   return 0;
 }
 
 static void serial_worker(void* data) {
   uv_barrier_t* barrier;
+  unsigned i;
 
   barrier = data;
+  for (i = 0; i < 5; i++)
+      uv_barrier_wait(barrier);
   if (uv_barrier_wait(barrier) > 0)
     uv_barrier_destroy(barrier);
 
@@ -123,16 +131,18 @@ TEST_IMPL(barrier_serial_thread) {
   uv_barrier_t barrier;
   unsigned i;
 
-  ASSERT(0 == uv_barrier_init(&barrier, ARRAY_SIZE(threads) + 1));
+  ASSERT_EQ(0, uv_barrier_init(&barrier, ARRAY_SIZE(threads) + 1));
 
   for (i = 0; i < ARRAY_SIZE(threads); ++i)
-    ASSERT(0 == uv_thread_create(&threads[i], serial_worker, &barrier));
+    ASSERT_EQ(0, uv_thread_create(&threads[i], serial_worker, &barrier));
 
+  for (i = 0; i < 5; i++)
+      uv_barrier_wait(&barrier);
   if (uv_barrier_wait(&barrier) > 0)
     uv_barrier_destroy(&barrier);
 
   for (i = 0; i < ARRAY_SIZE(threads); ++i)
-    ASSERT(0 == uv_thread_join(&threads[i]));
+    ASSERT_EQ(0, uv_thread_join(&threads[i]));
 
   return 0;
 }
@@ -141,8 +151,8 @@ TEST_IMPL(barrier_serial_thread) {
 TEST_IMPL(barrier_serial_thread_single) {
   uv_barrier_t barrier;
 
-  ASSERT(0 == uv_barrier_init(&barrier, 1));
-  ASSERT(0 < uv_barrier_wait(&barrier));
+  ASSERT_EQ(0, uv_barrier_init(&barrier, 1));
+  ASSERT_LT(0, uv_barrier_wait(&barrier));
   uv_barrier_destroy(&barrier);
   return 0;
 }
